+++ packages/mikro/src/lib/types.ts
@@ -1,5 +1,3 @@
-import type { AnyEntity } from '@mikro-orm/core';
-
 export interface MikroInitializerOptions {
-  serializeEntity: (entity: AnyEntity) => any;
+  unwrap: (value: any) => any;
 }